Moved files from bingmaps-v8 to bingmaps to clean up changes from V7 (deprecated)
