+++ bingmaps/Microsoft.Maps.Autosuggest.d.ts
