+++ bingmaps/Microsoft.Maps.Clustering.d.ts
