+++ bingmaps/Microsoft.Maps.DrawingTools.d.ts
