+++ bingmaps/Microsoft.Maps.GeoJSON.d.ts
