+++ bingmaps/Microsoft.Maps.HeatMapLayer.d.ts
