+++ bingmaps/Microsoft.Maps.SpatialDataService.d.ts
