+++ bingmaps/Microsoft.Maps.SpatialMath.d.ts
