+++ bingmaps/Microsoft.Maps.WellKnownText.d.ts
